Add `names` argument to ones, rand, randn, zeros, full; fix empty
In the short term, we implement this by having overloads for each of
these functions. In the long term, the plan is to move DimnameList to
TensorOptions so that we do not have to duplicate work.

Also fixes the implementation of empty. If there are no names, we should
just return an unnamed tensor instead of telling the user we don't
support their backend/layout.

Test Plan:
- [namedtensor ci]

gh-metadata: pytorch pytorch 24107 gh/zou3519/99/head